Things you should know. The database name is Project. You can change the name in the web.xml file. Database username and password are both root, you can change them in the Listener servlet. I added a third column to the users table like we did earlier but I called it USERID (bad choice, I know).
Also, all users get redirected to welcome.jsp when logging in, however, the page is displayed differently for admins, drivers and customers. I did this by setting a request attribute of 'admin' 'customer' or 'driver' in the servlet and used that variable to include either admin.html, driver.html or customer.html in the welcome.jsp page. 

Hope that makes sense.

Have fun!